+++ contracts/oracle/LPOracle.sol
@@ -16,11 +16,7 @@ contract LPOracle {
 
     /// @notice Constructor sets the address of the Nile AMM contract.
     /// @param _nileAMM The address of the Nile AMM contract.
-    constructor(
-        address _nileAMM,
-        address _zeroPriceFeed,
-        address _ethPriceFeed
-    ) {
+    constructor(address _nileAMM, address _zeroPriceFeed, address _ethPriceFeed) {
         nileAMM = INileAMM(_nileAMM);
         zeroPriceFeed = IPythAggregatorV3(_zeroPriceFeed);
         ethPriceFeed = IPythAggregatorV3(_ethPriceFeed);
@@ -30,15 +26,14 @@ contract LPOracle {
     /// @dev This function fetches reserves from the Nile AMM and uses a pre-defined price for tokens to calculate the LP token price.
     /// @return price The price of the liquidity pool token.
     function getPrice() public view returns (uint256 price) {
-        (uint256 reserve0, uint256 reserve1, ) = nileAMM.getReserves();
+        (uint256 reserve0, uint256 reserve1,) = nileAMM.getReserves();
 
         int256 px0 = zeroPriceFeed.latestAnswer();
         int256 px1 = ethPriceFeed.latestAnswer();
 
         require(px0 > 0 && px1 > 0, "Invalid Price");
 
-        uint256 sqrtK = (sqrt(reserve0 * reserve1) * 1e18) /
-            nileAMM.totalSupply();
+        uint256 sqrtK = (sqrt(reserve0 * reserve1) * 1e18) / nileAMM.totalSupply();
 
         price = (sqrtK * 2 * sqrt(uint256(px0 * px1))) / 1e18;
     }
@@ -47,9 +42,9 @@ contract LPOracle {
     /// @dev This function uses an iterative method to compute the square root of a number.
     /// @param x The number to compute the square root of.
     /// @return y The square root of the given number.
-    function sqrt(uint x) public pure returns (uint y) {
+    function sqrt(uint256 x) public pure returns (uint256 y) {
         if (x == 0) return 0; // Handle the edge case for 0
-        uint z = (x + 1) / 2;
+        uint256 z = (x + 1) / 2;
         y = x;
         while (z < y) {
             y = z;
